added us_amp and ds_amp to hdf5 output

+++ src/io/file/aps/aps_fit_params_import.cpp
@@ -450,7 +450,7 @@ bool APS_Fit_Params_Import::load(std::string path,
                     value.erase(std::remove(value.begin(), value.end(), '\n'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), '\r'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), ' '), value.end());
-                    params_override->us_amp_sens_num = std::stoi(value);
+                    params_override->us_amp_sens_num = std::stod(value);
                 }
                 else if (tag == "US_AMP_SENS_UNIT")
                 {
@@ -459,7 +459,7 @@ bool APS_Fit_Params_Import::load(std::string path,
                     value.erase(std::remove(value.begin(), value.end(), '\n'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), '\r'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), ' '), value.end());
-                    params_override->us_amp_sens_unit = std::stoi(value);
+                    params_override->us_amp_sens_unit = std::stod(value);
                 }
                 else if (tag == "DS_AMP_SENS_NUM")
                 {
@@ -468,7 +468,7 @@ bool APS_Fit_Params_Import::load(std::string path,
                     value.erase(std::remove(value.begin(), value.end(), '\n'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), '\r'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), ' '), value.end());
-                    params_override->ds_amp_sens_num = std::stoi(value);
+                    params_override->ds_amp_sens_num = std::stod(value);
                 }
                 else if (tag == "DS_AMP_SENS_UNIT")
                 {
@@ -477,7 +477,7 @@ bool APS_Fit_Params_Import::load(std::string path,
                     value.erase(std::remove(value.begin(), value.end(), '\n'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), '\r'), value.end());
                     value.erase(std::remove(value.begin(), value.end(), ' '), value.end());
-                    params_override->ds_amp_sens_unit = std::stoi(value);
+                    params_override->ds_amp_sens_unit = std::stod(value);
                 }
                 else if (tag == "BE_WINDOW_THICKNESS")
                 {
